requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
dataclasses==0.8
decorator==5.0.9
discord==1.7.3
discord.py==1.7.3
docopt==0.6.2
et-xmlfile==1.1.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.16
grpcio==1.34.1
h5py==3.1.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.6.0
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.2.0
jupyter-core==4.7.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
mne==0.23.0
multidict==5.1.0
nest-asyncio==1.5.1
numexpr==2.7.3
numpy==1.19.5
oauthlib==3.1.1
openpyxl==3.0.7
opt-einsum==3.3.0
pandas==1.1.5
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
pipreqs==0.4.10
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pydot-ng==2.0.0
pydotplus==2.0.2
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
sklearn==0.0
tables==3.6.1
tdqm==0.0.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
threadpoolctl==2.2.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.61.2
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
XlsxWriter==3.0.1
yarg==0.1.9
yarl==1.6.3
youtube-dl==2021.6.6
zipp==3.4.1
